refactor: split internal/app/input.go by responsibility
Closes #93

- navigation.go: cursor movement, scroll, PR open actions
- review_input.go: review key handling and diff-mode guards
- filter_input.go: filter menu key handling
- input.go: key dispatch and global actions only
