+++ internal/app/input.go
@@ -7,29 +7,6 @@ import (
 	"github.com/rin2yh/lazygh/internal/model"
 )
 
-func (s *screen) handleReviewInputKey(msg tea.KeyMsg) (tea.Cmd, bool) {
-	action, ok := s.gui.config.KeyBindings.ActionFor(msg)
-	if ok {
-		switch action {
-		case config.ActionReviewSubmit:
-			return s.gui.review.Submit(), true
-		case config.ActionReviewDiscard:
-			return s.gui.review.Discard(), true
-		case config.ActionReviewSave:
-			if s.gui.review.InputMode() == model.ReviewInputComment {
-				if s.gui.review.IsEditingComment() {
-					return s.gui.review.SaveEditComment(), true
-				}
-				return s.gui.review.SaveComment(), true
-			}
-		}
-	}
-	if cmd, handled := s.gui.review.EditorKey(msg); handled {
-		return cmd, true
-	}
-	return nil, false
-}
-
 func (s *screen) handleKeyInput(msg tea.KeyMsg) tea.Cmd {
 	action, ok := s.gui.config.KeyBindings.ActionFor(msg)
 	if !ok {
@@ -78,73 +55,6 @@ func (s *screen) handleGlobalAction(action config.Action) (tea.Cmd, bool) {
 	}
 }
 
-func (s *screen) handleNavigationAction(action config.Action) (tea.Cmd, bool) {
-	switch action {
-	case config.ActionMoveDown:
-		return s.moveDown(), true
-	case config.ActionMoveUp:
-		return s.moveUp(), true
-	case config.ActionPageDown, config.ActionPageUp, config.ActionGoTop, config.ActionGoBottom:
-		return s.handleDetailScrollAction(action), true
-	default:
-		return nil, false
-	}
-}
-
-func (s *screen) handleReviewAction(action config.Action) tea.Cmd {
-	switch action {
-	case config.ActionReviewRange:
-		return s.startReviewRange()
-	case config.ActionReviewComment:
-		return s.startReviewComment()
-	case config.ActionReviewSummary:
-		return s.startReviewSummary()
-	case config.ActionReviewSubmit:
-		return s.gui.review.Submit()
-	case config.ActionReviewDiscard:
-		return s.gui.review.Discard()
-	case config.ActionReviewClearComment:
-		if s.gui.review.InputMode() == model.ReviewInputComment {
-			s.gui.review.ClearCommentInput()
-		}
-	case config.ActionReviewEvent:
-		if s.gui.coord.IsDiffMode() {
-			s.gui.review.CycleReviewEvent()
-		}
-	case config.ActionReviewDeleteComment:
-		if s.gui.focus == layout.FocusReviewDrawer {
-			return s.gui.review.DeleteComment()
-		}
-	case config.ActionReviewEditComment:
-		if s.gui.focus == layout.FocusReviewDrawer {
-			s.gui.review.EditComment()
-		}
-	}
-	return nil
-}
-
-func (s *screen) handleFilterKey(msg tea.KeyMsg) tea.Cmd {
-	switch msg.String() {
-	case "esc":
-		s.gui.coord.CloseFilterSelect()
-		return nil
-	case "enter":
-		s.gui.coord.CloseFilterSelect()
-		s.gui.coord.BeginFetchPRs()
-		return s.loadPRsCmd()
-	case "j", "down":
-		s.gui.coord.MoveFilterCursor(1)
-		return nil
-	case "k", "up":
-		s.gui.coord.MoveFilterCursor(-1)
-		return nil
-	case " ":
-		s.gui.coord.ToggleFilterAtCursor()
-		return nil
-	}
-	return nil
-}
-
 func (s *screen) handleCancel() tea.Cmd {
 	if s.gui.review.InputMode() == model.ReviewInputNone && s.gui.review.HasRangeStart() {
 		s.gui.review.ClearRangeStart()
@@ -159,162 +69,4 @@ func (s *screen) handleCancel() tea.Cmd {
 	}
 	s.gui.focusPRs()
 	return nil
-}
-
-func (s *screen) showDiff() tea.Cmd {
-	if s.gui.switchToDiff() {
-		return s.openSelectedPR()
-	}
-	return nil
-}
-
-func (s *screen) moveDown() tea.Cmd { return s.moveCursor(1) }
-func (s *screen) moveUp() tea.Cmd   { return s.moveCursor(-1) }
-
-// moveCursor moves the cursor in the given direction (1 = down, -1 = up).
-func (s *screen) moveCursor(dir int) tea.Cmd {
-	if s.gui.coord.IsDiffMode() {
-		switch s.gui.focus {
-		case layout.FocusPRs:
-			navigate := s.gui.navigateDown
-			if dir < 0 {
-				navigate = s.gui.navigateUp
-			}
-			if navigate() {
-				return s.openSelectedPR()
-			}
-			return nil
-		case layout.FocusDiffFiles:
-			if dir > 0 {
-				s.gui.diff.SelectNextFile()
-			} else {
-				s.gui.diff.SelectPrevFile()
-			}
-			return nil
-		case layout.FocusDiffContent:
-			if dir > 0 {
-				s.scrollDetailDown()
-			} else {
-				s.scrollDetailUp()
-			}
-			return nil
-		case layout.FocusReviewDrawer:
-			if dir > 0 {
-				s.gui.review.SelectNextComment()
-			} else {
-				s.gui.review.SelectPrevComment()
-			}
-			return nil
-		}
-		return nil
-	}
-
-	if s.gui.focus == layout.FocusPRs {
-		if dir > 0 {
-			s.gui.navigateDown()
-		} else {
-			s.gui.navigateUp()
-		}
-	}
-	return nil
-}
-
-func (s *screen) openSelectedPR() tea.Cmd {
-	action := s.gui.coord.PlanEnter(s.gui.client != nil)
-	switch action.Kind {
-	case model.EnterLoadPRDiff:
-		return s.loadDetailCmd(action.Repo, action.Number, model.DetailModeDiff)
-	case model.EnterLoadPRDetail:
-		return s.loadDetailCmd(action.Repo, action.Number, model.DetailModeOverview)
-	default:
-		return nil
-	}
-}
-
-func (s *screen) handleDetailScrollAction(action config.Action) tea.Cmd {
-	if s.gui.focus != layout.FocusDiffContent {
-		return nil
-	}
-
-	if s.gui.coord.IsDiffMode() {
-		switch action {
-		case config.ActionPageUp:
-			s.gui.diff.SelectPrevLine(s.gui.detail.Height())
-		case config.ActionPageDown:
-			s.gui.diff.SelectNextLine(s.gui.detail.Height())
-		case config.ActionGoTop:
-			s.gui.diff.GotoFirstLine()
-		case config.ActionGoBottom:
-			s.gui.diff.GotoLastLine()
-		}
-		return nil
-	}
-
-	switch action {
-	case config.ActionPageDown, config.ActionPageUp:
-		key, ok := primaryKeyMsg(s.gui.config.KeyBindings.Binding(action))
-		if !ok {
-			return nil
-		}
-		_, cmd := s.gui.detail.Update(key)
-		return cmd
-	}
-	return nil
-}
-
-func primaryKeyMsg(binding config.KeyBinding) (tea.KeyMsg, bool) {
-	if len(binding.Keys) == 0 {
-		return tea.KeyMsg{}, false
-	}
-	switch binding.Keys[0] {
-	case "pgdown":
-		return tea.KeyMsg{Type: tea.KeyPgDown}, true
-	case "pgup":
-		return tea.KeyMsg{Type: tea.KeyPgUp}, true
-	case " ":
-		return tea.KeyMsg{Type: tea.KeySpace}, true
-	case "b":
-		return tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'b'}}, true
-	case "f":
-		return tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'f'}}, true
-	default:
-		return tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune(binding.Keys[0])}, true
-	}
-}
-
-func (s *screen) scrollDetailDown() {
-	if s.gui.coord.IsDiffMode() {
-		s.gui.diff.SelectNextLine(1)
-		return
-	}
-	s.gui.detail.ScrollDown(1)
-}
-
-func (s *screen) scrollDetailUp() {
-	if s.gui.coord.IsDiffMode() {
-		s.gui.diff.SelectPrevLine(1)
-		return
-	}
-	s.gui.detail.ScrollUp(1)
-}
-
-func (s *screen) requireDiffMode(notice string, fn func()) tea.Cmd {
-	if !s.gui.coord.IsDiffMode() {
-		s.gui.review.SetNotice(notice)
-		return nil
-	}
-	fn()
-	return nil
-}
-
-func (s *screen) startReviewRange() tea.Cmd {
-	return s.requireDiffMode("Review range selection is only available in diff view.", s.gui.review.ToggleRangeSelection)
-}
-
-func (s *screen) startReviewComment() tea.Cmd {
-	return s.requireDiffMode("Review comments are only available in diff view.", s.gui.review.BeginCommentFlow)
-}
-
-func (s *screen) startReviewSummary() tea.Cmd {
-	return s.requireDiffMode("Review summary is only available in diff view.", s.gui.review.BeginSummaryInput)
 }
